backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
Pillow==11.1.0
razorpay==1.4.2
